renaming to colormode, default panicing
