+++ internal/colormode/detect_linux.go
@@ -1,20 +1,20 @@
-//go:build linux
-// +build linux
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: 2025 The Guigui Authors
 
-package theme
+package colormode
 
 import (
 	"os/exec"
 	"strings"
 )
 
-func detectSystemTheme() ColorMode {
+func systemColorMode() ColorMode {
 	out, err := exec.Command("gsettings", "get", "org.gnome.desktop.interface", "gtk-theme").Output()
 	if err != nil {
-		return ThemeUnknown
+		return Unknown
 	}
 	if strings.Contains(strings.ToLower(string(out)), "dark") {
-		return ThemeDark
+		return Dark
 	}
-	return ThemeLight
+	return Light
 }
