+++ internal/colormode/detect_windows.go
@@ -1,28 +1,28 @@
-//go:build windows
-// +build windows
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: 2025 The Guigui Authors
 
-package theme
+package colormode
 
 import (
 	"golang.org/x/sys/windows/registry"
 )
 
-func detectSystemTheme() ColorMode {
+func systemColorMode() ColorMode {
 	k, err := registry.OpenKey(registry.CURRENT_USER,
 		`Software\Microsoft\Windows\CurrentVersion\Themes\Personalize`,
 		registry.QUERY_VALUE)
 	if err != nil {
-		return ThemeUnknown
+		return Unknown
 	}
 	defer k.Close()
 
 	val, _, err := k.GetIntegerValue("AppsUseLightTheme")
 	if err != nil {
-		return ThemeUnknown
+		return Unknown
 	}
 
 	if val == 0 {
-		return ThemeDark
+		return Dark
 	}
-	return ThemeLight
+	return Light
 }
